feat: create address update and delete operation

+++ Customer.Datalayer/src/Customer.Datalayer.WebForm/CustomersList.aspx.cs
@@ -32,13 +32,12 @@ namespace Customer.Datalayer.WebForm
         protected void Page_Load(object sender, EventArgs e)
         {
             LoadCustomersFromDatabase();
-            
-        }
+            if (!IsPostBack)
+            {
+                var customerIDReq = Convert.ToInt32(Request.QueryString["customerID"]);
+                _customerRepository.Delete(customerIDReq);
+            }
 
-        protected void OnClickDelete(object sender, EventArgs e)
-        {
-            var customerIDReq = Convert.ToInt32(Request.QueryString["customerID"]);
-            _customerRepository.Delete(customerIDReq);
         }
     }
 }
